pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asyncdb"
description = "Asynchronous database package."
dynamic = ["version"]
requires-python = ">=3.10"
authors = [
    { name = "Matimu Chauke", email = "matimu@chauke.org" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "asyncpg==0.29.0",
    "greenlet==3.0.3",
    "sqlmodel==0.0.14",
    "alembic==1.13.1",
    "SQLAlchemy==2.0.25",
    "pydantic-settings==2.1.0"
]

[project.optional-dependencies]
full = []

[tool.hatch.version]
path = "asyncdb/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/asyncdb",
]
